split the SDL music backend into a separate source file
