+++ Source/i_sound.c
@@ -28,20 +28,16 @@
 
 // haleyjd
 #include "SDL.h"
-#include "SDL_audio.h"
 #include "SDL_mixer.h"
 #include <math.h>
 
+#include "i_sdlmusic.h"
 #include "i_oplmusic.h"
 
-#include "z_zone.h"
 #include "doomstat.h"
-#include "mmus2mid.h"   //jff 1/16/98 declarations for MUS->MIDI converter
 #include "i_sound.h"
 #include "w_wad.h"
-#include "g_game.h"     //jff 1/21/98 added to use dprintf in I_RegisterSong
 #include "d_main.h"
-#include "d_io.h"
 
 // Needed for calling the actual sound output.
 int SAMPLECOUNT = 512;
@@ -68,9 +64,6 @@ int default_mus_card;
 // these routines think that sound has been initialized when it hasn't
 boolean snd_init = false;
 boolean mus_init = false;
-
-int detect_voices; //jff 3/4/98 enables voice detection prior to install_sound
-//jff 1/22/98 make these visible here to disable sound/music on install err
 
 // haleyjd 10/28/05: updated for Julian's music code, need full quality now
 int snd_samplerate = 44100;
@@ -689,9 +682,13 @@ void I_ShutdownSound(void)
 //
 // SoM 9/14/02: Rewrite. code taken from prboom to use SDL_Mixer
 //
-static void I_InitMusicBackend();
 void I_InitSound(void)
 {   
+   if (0)
+      I_SDL_InitMusicBackend();
+   else
+      I_OPL_InitMusicBackend();
+
    if(!nosfxparm)
    {
       int audio_buffers;
@@ -750,38 +747,9 @@ void I_InitSound(void)
       // haleyjd 04/11/03: don't use music if sfx aren't init'd
       // (may be dependent, docs are unclear)
       if(!nomusicparm)
-      {
-         I_InitMusicBackend();
          I_InitMusic();
-      }
    }   
 }
-
-///
-// MUSIC API.
-//
-
-// julian (10/25/2005): rewrote (nearly) entirely
-
-#include "mmus2mid.h"
-#include "m_misc.h"
-#include "m_misc2.h"
-#include "i_midipipe.h"
-
-// Only one track at a time
-static Mix_Music *music = NULL;
-
-// Some tracks are directly streamed from the RWops;
-// we need to free them in the end
-static SDL_RWops *rw = NULL;
-
-// Same goes for buffers that were allocated to convert music;
-// since this concerns mus, we could do otherwise but this 
-// approach is better for consistency
-static void *music_block = NULL;
-
-// Macro to make code more readable
-#define CHECK_MUSIC(h) ((h) && music != NULL)
 
 boolean (*I_InitMusic)(void);
 void (*I_ShutdownMusic)(void);
@@ -792,364 +760,3 @@ void *(*I_RegisterSong)(void *data, int size);
 void (*I_PlaySong)(void *handle, boolean looping);
 void (*I_StopSong)(void *handle);
 void (*I_UnRegisterSong)(void *handle);
-
-//
-// I_ShutdownMusic
-//
-// atexit handler.
-//
-static void I_SDL_UnRegisterSong(void *handle);
-static void I_SDL_ShutdownMusic(void)
-{
-#if defined(_WIN32)
-   if (midi_server_initialized)
-   {
-      I_MidiPipe_ShutdownServer();
-   }
-   else
-#endif
-   {
-      I_SDL_UnRegisterSong((void *)1);
-   }
-}
-
-//
-// I_InitMusic
-//
-static boolean I_SDL_InitMusic(void)
-{
-   switch(mus_card)
-   {
-   case -1:
-      printf("I_InitMusic: Using SDL_mixer.\n");
-      mus_init = true;
-
-      // Initialize SDL_Mixer for MIDI music playback
-      Mix_Init(MIX_INIT_MID | MIX_INIT_FLAC | MIX_INIT_OGG | MIX_INIT_MP3); // [crispy] initialize some more audio formats
-   #if defined(_WIN32)
-      // [AM] Start up midiproc to handle playing MIDI music.
-      I_MidiPipe_InitServer();
-   #endif
-      break;   
-   default:
-      printf("I_InitMusic: Music is disabled.\n");
-      break;
-   }
-   
-   atexit(I_SDL_ShutdownMusic);
-
-   return mus_init;
-}
-
-// jff 1/18/98 changed interface to make mididata destroyable
-
-static void I_SDL_SetMusicVolume(int volume);
-static void I_SDL_PlaySong(void *handle, boolean looping)
-{
-   if(!mus_init)
-      return;
-
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_PlaySong(looping ? -1 : 1);
-   }
-   else
-#endif
-   if(CHECK_MUSIC(handle) && Mix_PlayMusic(music, looping ? -1 : 0) == -1)
-   {
-      dprintf("I_PlaySong: Mix_PlayMusic failed\n");
-      return;
-   }
-   
-   // haleyjd 10/28/05: make sure volume settings remain consistent
-   I_SDL_SetMusicVolume(snd_MusicVolume);
-}
-
-//
-// I_SetMusicVolume
-//
-
-static int current_midi_volume;
-
-static void I_SDL_SetMusicVolume(int volume)
-{
-   // haleyjd 09/04/06: adjust to use scale from 0 to 15
-   current_midi_volume = (volume * 128) / 15;
-
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_SetVolume(current_midi_volume);
-   }
-   else
-#endif
-   {
-      Mix_VolumeMusic(current_midi_volume);
-   }
-}
-
-//
-// I_PauseSong
-//
-static void I_SDL_PauseSong(void *handle)
-{
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_SetVolume(0);
-   }
-   else
-#endif
-   if(CHECK_MUSIC(handle))
-   {
-      // Not for mids
-      if(Mix_GetMusicType(music) != MUS_MID)
-         Mix_PauseMusic();
-      else
-      {
-         // haleyjd 03/21/06: set MIDI volume to zero on pause
-         Mix_VolumeMusic(0);
-      }
-   }
-}
-
-//
-// I_ResumeSong
-//
-static void I_SDL_ResumeSong(void *handle)
-{
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_SetVolume(current_midi_volume);
-   }
-   else
-#endif
-   if(CHECK_MUSIC(handle))
-   {
-      // Not for mids
-      if(Mix_GetMusicType(music) != MUS_MID)
-         Mix_ResumeMusic();
-      else
-         Mix_VolumeMusic(current_midi_volume);
-   }
-}
-
-//
-// I_StopSong
-//
-static void I_SDL_StopSong(void *handle)
-{
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_StopSong();
-   }
-   else
-#endif
-   if(CHECK_MUSIC(handle))
-      Mix_HaltMusic();
-}
-
-//
-// I_UnRegisterSong
-//
-static void I_SDL_UnRegisterSong(void *handle)
-{
-#if defined(_WIN32)
-   if (midi_server_registered)
-   {
-      I_MidiPipe_UnregisterSong();
-   }
-   else
-#endif
-   if(CHECK_MUSIC(handle))
-   {   
-      // Stop and free song
-      I_SDL_StopSong(handle);
-      Mix_FreeMusic(music);
-      
-      // Free RWops
-      if(rw != NULL)
-         SDL_FreeRW(rw);
-      
-      // Free music block
-      if(music_block != NULL)
-         free(music_block);
-      
-      // Reinitialize all this
-      music = NULL;
-      rw = NULL;
-      music_block = NULL;
-   }
-}
-
-#if defined(_WIN32)
-static void MidiProc_RegisterSong(void *data, int size)
-{
-   char* filename;
-   filename = M_TempFile("doom"); // [crispy] generic filename
-
-   M_WriteFile(filename, data, size);
-
-   if (!I_MidiPipe_RegisterSong(filename))
-   {
-      fprintf(stderr, "Error loading midi: %s\n",
-          "Could not communicate with midiproc.");
-   }
-   (free)(filename);
-}
-#endif
-
-//
-// I_RegisterSong
-//
-static void *I_SDL_RegisterSong(void *data, int size)
-{
-   if(music != NULL)
-      I_SDL_UnRegisterSong((void *)1);
-
-   if (size < 4 || memcmp(data, "MUS\x1a", 4)) // [crispy] MUS_HEADER_MAGIC
-   {
-   #if defined(_WIN32)
-      if (size >= 4 && memcmp(data, "MThd", 4) == 0 && midi_server_initialized) // MIDI header magic
-      {
-         music = NULL;
-         MidiProc_RegisterSong(data, size);
-         return (void *)1;
-      }
-      else
-   #endif
-      {
-         rw    = SDL_RWFromMem(data, size);
-         music = Mix_LoadMUS_RW(rw, false);
-      }
-   }
-   else // Assume a MUS file and try to convert
-   {
-      int err;
-      MIDI mididata;
-      UBYTE *mid;
-      int midlen;
-
-      memset(&mididata, 0, sizeof(MIDI));
-
-      if((err = mmus2mid((byte *)data, &mididata, 89, 0)))
-      {
-         // Nope, not a mus.
-         dprintf("Error loading music: %d", err);
-         return NULL;
-      }
-
-      // Hurrah! Let's make it a mid and give it to SDL_mixer
-      MIDIToMidi(&mididata, &mid, &midlen);
-
-   #if defined(_WIN32)
-      if (midi_server_initialized)
-      {
-         music = NULL;
-         MidiProc_RegisterSong(mid, midlen);
-         free(mid);
-         return (void *)1;
-      }
-      else
-   #endif
-      {
-         rw    = SDL_RWFromMem(mid, midlen);
-         music = Mix_LoadMUS_RW(rw, false);
-
-         if(music == NULL)
-         {
-            // Conversion failed, free everything
-            SDL_FreeRW(rw);
-            rw = NULL;
-            free(mid);
-         }
-         else
-         {
-            // Conversion succeeded
-            // -> save memory block to free when unregistering
-            music_block = mid;
-         }
-      }
-   }
-   
-   // the handle is a simple boolean
-   return music;
-}
-
-static void I_InitMusicBackend()
-{
-	if (1)
-	{
-		I_InitMusic = I_OPL_InitMusic;
-		I_ShutdownMusic = I_OPL_ShutdownMusic;
-		I_SetMusicVolume = I_OPL_SetMusicVolume;
-		I_PauseSong = I_OPL_PauseSong;
-		I_ResumeSong = I_OPL_ResumeSong;
-		I_RegisterSong = I_OPL_RegisterSong;
-		I_PlaySong = I_OPL_PlaySong;
-		I_StopSong = I_OPL_StopSong;
-		I_UnRegisterSong = I_OPL_UnRegisterSong;
-	}
-	else
-	{
-		I_InitMusic = I_SDL_InitMusic;
-		I_ShutdownMusic = I_SDL_ShutdownMusic;
-		I_SetMusicVolume = I_SDL_SetMusicVolume;
-		I_PauseSong = I_SDL_PauseSong;
-		I_ResumeSong = I_SDL_ResumeSong;
-		I_RegisterSong = I_SDL_RegisterSong;
-		I_PlaySong = I_SDL_PlaySong;
-		I_StopSong = I_SDL_StopSong;
-		I_UnRegisterSong = I_SDL_UnRegisterSong;
-	}
-}
-
-//----------------------------------------------------------------------------
-//
-// $Log: i_sound.c,v $
-// Revision 1.15  1998/05/03  22:32:33  killough
-// beautification, use new headers/decls
-//
-// Revision 1.14  1998/03/09  07:11:29  killough
-// Lock sound sample data
-//
-// Revision 1.13  1998/03/05  00:58:46  jim
-// fixed autodetect not allowed in allegro detect routines
-//
-// Revision 1.12  1998/03/04  11:51:37  jim
-// Detect voices in sound init
-//
-// Revision 1.11  1998/03/02  11:30:09  killough
-// Make missing sound lumps non-fatal
-//
-// Revision 1.10  1998/02/23  04:26:44  killough
-// Add variable pitched sound support
-//
-// Revision 1.9  1998/02/09  02:59:51  killough
-// Add sound sample locks
-//
-// Revision 1.8  1998/02/08  15:15:51  jim
-// Added native midi support
-//
-// Revision 1.7  1998/01/26  19:23:27  phares
-// First rev with no ^Ms
-//
-// Revision 1.6  1998/01/23  02:43:07  jim
-// Fixed failure to not register I_ShutdownSound with atexit on install_sound error
-//
-// Revision 1.4  1998/01/23  00:29:12  killough
-// Fix SSG reload by using frequency stored in lump
-//
-// Revision 1.3  1998/01/22  05:55:12  killough
-// Removed dead past changes, changed destroy_sample to stop_sample
-//
-// Revision 1.2  1998/01/21  16:56:18  jim
-// Music fixed, defaults for cards added
-//
-// Revision 1.1.1.1  1998/01/19  14:02:57  rand
-// Lee's Jan 19 sources
-//
-//----------------------------------------------------------------------------
